Move visitor classes according to PSR-4
