+++ composer.json
@@ -26,6 +26,7 @@
     "minimum-stability": "stable",
     "autoload-dev": {
         "classmap": [
+            "src/",
             "tests/"
         ]
     },
