+++ phpstan.neon.dist
@@ -3,6 +3,7 @@ parameters:
         - finder.php
         - functionMap.php
         - visitor.php
+        - src/
         - tests/
     excludePaths:
         - tests/data/
